pyproject.toml
[project]
name = "sql-optimizer-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "mcp[cli]>=1.0.0",
    "mysql-connector-python>=9.6.0",
    "psycopg[binary,pool]>=3.1",
    "python-dotenv>=1.2.2",
    "sqlglot>=23.0",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]
